I15 (9.16) Add suppressOpenEHShutterAtScanStart support
This mirrors suppressCloseEHShutterAtScanEnd just as
exposeSuppressOpenDetectorShieldAtScanStart mirrors
exposeSuppressCloseDetectorShieldAtScanEnd.

This was useful during 2020 run 2, when we had no control over the
Experimental Hutch shutter.

Also, don't try to close the Detector Shield on error or stop if
suppressCloseDetectorShieldAtScanEnd is defined.

This was useful during 2020 run 2, when attempting to close an already
closed Experimental Hutch shutter resulted in errors being reported from
EPICS.

Change-Id: I86d81789a721bbd6dc2ec2c5ce07b8de704bdaa5