DAQ-3016 Delete symlinks to dummy products
Only build live products with Tycho (as Buckminster does). Tycho pomless
doesn't currently support multiple products per project
(https://bugs.eclipse.org/bugs/show_bug.cgi?id=562887) so it is
indeterminate which product will be built.

Note that for b16 and i16 the '.live.' version of the product is not
the product that is actually deployed.

Change-Id: Ib65d053b3f7df5c94e34f80b35c0414f89237305