DAQ-3016 Delete symlinks to dummy products
Only build live products with Tycho (as Buckminster does). Tycho pomless
doesn't currently support multiple products per project
(https://bugs.eclipse.org/bugs/show_bug.cgi?id=562887) so it is
indeterminate which product will be built.

Note that for i15 and i15-1 the '.live.' version of the product is not
the product that is actually deployed.

Change-Id: I30ce907613dcd3906a60c383105c3a075c43072c