DAQ-2618 Add appender to log Jython loaded classes to separate file
Using GDAJythonClassloader log the Java classes successfully loaded
by Jython.

A file based appender is used with a filename including the timestamp
of the context creation. This causes a new file to be created each time
the server is started. Different locations are defined (via properties)
for dummy and live modes.

This is to assist with technical debt reporting. A processing script/
Jenkins job will be responsible for interpreting the results and clean
up of old files.

Change-Id: I29d402df4a57bf034c2ef07df86039a66caedf46