Enable Tycho tests for modules in dls-beamlines-xas
Some issues here are related to split packages.

In Tycho it appears that accessing package private members from a class
in the same package but different plugin is not allowed and gives an
IllegalAccessError. Not sure if this is also true in our
GDA Server/Client runtime.

Therefore cannot call scan.callScannablesAtScanStart() since it's a
protected method.

Similarly the FastShutter enum had package private members so replace
with public getter methods.

Change-Id: If5e52644563bd63a30240881ca839c8cf5156d95