K11-819 Differentiate between live and dummy properties
This is necessary because Malcolm simulator defines scannable like
sim_x, sim_y, sim_z, simtheta while the
servers/main/_common/stages/gts.xml, using <gda:motor>, both dummy and
live scannables with the same name, i.e. gts_theta.
Consequently when the scanRequest is assembled the scannable in GDA is
named gda_theta, while the scannable in malcolm-tomo, simulator, is
simtheta.
The properties defined in the client, define the stage, but also help to
mitigate this instructing the client about which are the motors to use
for a tomography scan.

Change-Id: Id58212ba77b901fc639d4eefecd53b32d340aafb