I20_1-321 : Make 'live stream view' for Ede detectors (Frelon/XH)
A new live stream custom UI has been made - this has controls to start,
stop the detector and adjust the acquisition parameters. When the
detector is started, 2D image data from the detector is sent to the
live stream plot via the ImageDatasetConnector.
Updates to EdeFrelon :
- New functions added to allow ROI mode to be set; 'kinetic ROI' mode is
automatically set at the start of scans.
- Detector now includes vertical binning and vertical start parameters -
these are the defaults used when setting 'kinetic ROI' mode.
- Exception handling when setting up the detector has been refactored
and simplified.

Change-Id: I72d3c450fe648138797067cb4e9ac943fc0950ad