DAQ-3068 Make XasScanBase#createScanArguments public
This is method is called outside this class in test
methods.

It was a protected method which allows package private access too
however due to the split packages of the test plugins this access
will only work if the test plugins are all fragments of
uk.ac.gda.server.exafs which they are not.

In OSGi package private doesn't work across the class loader boundary.

Change-Id: I69073bb78db0c75f751c7087adbd0699689d1e66