I14-829 Fix I14 BeamlineReadinessDisplay to handle energy out of range
One of the criteria of beamline readiness is whether the beam
is at its target intensity. The target intensity depends on the
DCM energy and is calculated by interpolation from a series of values
provided by the beamline.

There has been a case where the energy was outside the provided series
of values: in this case the interpolation function throws an exception
which was not being caught and the status view was not being properly
created.

This change detects an energy out of range and handles it and shows the
beam status as yellow ("intensity low")

NB A better fix is available for GDA 9.20 by deriving this
class from FourStateDisplay and showing the beam state as "unknown", but
FourStateDisplay is not available in GDA 9.18 and adding it would also
require more changes than we should make to a release branch.

Change-Id: I74e23cb2986b89e7036266c4adc1a8acd22b5c74