DAQ-3214 Delete obsolete site feature projects
Under Buckminster these were used to generate the correct p2 for each
product and export the products from this. The Buckminster product build
scripts were also stored here along with beamline/Jenkins specific
properties.

For the Tycho build we have created new product projects (which are
plain projects rather than features) as this is the pattern supported by
Tycho.

In general the site suffix would usually refer to an update site however
none of the GDA site projects actually define an update site (in
contrast to Dawn for example).

If we require update sites in the future we could create new site
projects but there is no current use case for this.

Change-Id: I265b939c3fcab2302f1f5f5cbb4b5ea267f90403