DAQ-3214 Delete all-dls-{clients,configs} aggregator projects
These were previously used by Buckminster as top level materialize
requirements which had cspec requirements on all the individual beamline
projects.

With Oomph we are currently maintaining the equivalent of this in the
setup file itself for client features/products. Config projects are
now Oomph dependencies of each beamline's feature/product project.

It would be possible to use the aggregator project approach with Oomph
as well but there are pros and cons to both. The all-dls-clients would
have to be restructured anyway as it currently refers to site projects.

Change-Id: I6a665b7aad31b617d55ac06d7d6212a697b49b93