DAQ-3310 Remove product icons
The icons cause issues with the Tycho build as each project is built
in its own Maven module context so paths referring to other Eclipse
projects are only valid when there is an Eclipse workspace present.
For details see https://bugs.eclipse.org/bugs/show_bug.cgi?id=424805.

A product icon gets placed in the root of the product as icon.xpm (the
name is platform specific). Note that on Linux executables don't have
icons associated with them without creating a desktop entry (.desktop
file).

For GDA server and clients we don't actually require icons here as the
products are launched from the terminal or the Diamond launcher (which
has its own source of icons). There are some .desktop files in various
configs but they source icons from elsewhere.

Note that the product icon differs from the other branding elements such
as window images which, although they look to be defined with similar
paths, actually are synchronised to extension points in a plugin so
still work with Tycho.

If we really decide that we need icons in future then they should be
added in a Tycho compatible way, the cleanest approach is to put a copy
in the same directory as the product file.

Change-Id: Iac7801d58ca67f1a00832c7e6b1c6a03cda3b3de