I20_1-329 : Imrovements for Ede scan data collection
- Ensure all available frames are read out out from the Frelon at the
end of each collection. (lastImageReady was not always the set to the
last available frame, and sometimes the last image in a scan/cycle would
be missing).
- Any 'Interrupted Exceptions' thrown when running a scan are now
caught, so that post-scan Nexus file processing can be done on the
collected data.
- 'Smart stop' now works for Cyclic experiments.
- Minor refactor to eliminate the need to store the current timing group
as a member variable in EdeScan. It is instead looked up from the
detector object.

Change-Id: I4775bcc1a10b0721004a60620ed58c64696eae34