set ChildFieldNode sourceFieldName to "value"
instead of org.eclipse.dawnsci.nexus.NXpositioner.NX_VALUE. The property
value is interpreted as a literal string and not evaluted to the value
of that constant. I don't know if it is possible in spring to refer
to a Java constant. This only worked as in fact the value wasn't read,
CompositeNexusScannable always used "value".

Change-Id: If2b5de3132b9eefd1967f498137d5465b023b64d