I07-348 SCI-9560 Enable new DESELECT_OTHERS for DataVis
This mode matches behaiour of old Scan Polt view such that when a new
scan file is loaded it is automatically selected and other open files
are deselected.

This should become an Eclipse Preference eventually so that we don't
need to modify the product file. Alternatively make it possible to set
from LocalProperties.

Change-Id: Ie6017cecc6f81d007cb07010353d44e726d4be86