I07-348 Create new I07 Scan perspective
This is a comprised of views from Dawn's DataVis perspective and
other common GDA views required for scripting and scanning.

The default File menu has been hidden in this perspective and replaced
by a custom menu.

Includes fix for I07-347: Set preference to disable opening old plot
in new perspective.

Change-Id: I078777ceee8222d085b9db11de4eebfe51a2e511