I18: Refactor EarlyStartup
This historic class did several things, most of which (if not all) are
no longer needed. In addition, it was the source of two recent obscure
bugs: NPE the first time one manually switched from the plot perspective
to the experiment perspective; and the empty editor area which should be
hidden in the plotting perspective being made visible at scan start.

Refactoring this class to keep just one of the previous functions has
resolved these issues.

A summary of what the class used to do:

- Some overhead to enable configuring custom perspective from an XML
  representation generated by Eclipse at runtime:
  No longer used, so removed.
- Opening default perspectives based on properties in customization file
  with a note "This shouldn't be necessary":
  Not necessary, so removed.
- Manage the opening and closing of the experiment perspective editors:
  This is already handled in the perspective itself, so removed.
- Cache and restore editor selection when switching perspectives:
  This didn't work - the perspectiveDeactivated method was not called as
  expected.
- Refreshed ExafsSelectionView when switching to Microfocus perspective:
  Perspective not used (will be deleted soon), so removed.

The only thing left in the class is a listener to refresh the XAFS
project when the ExperimentPerspective is switched to. It may be that
this is no longer necessary but it needs more investigation.

Change-Id: Ie8e428d1fc0055fe19cd524de8834f3c5671ea57