i18: Remove needless early return for dummy mode in CameraView
The early return means that no part of the view is created in dummy mode
and there is no reason for this:

The live stream will be available. The start/stop actions are EPICS
operations so they will fail in dummy mode due to network configuration.

The return is now removed.


Change-Id: If180a2ffd599e6757cc05c3d665b44960e664ef1