K11-717 Move acquisition to click coordinates in Point and Shoot
The handler was missing the <move acquisition to click coordinates> part
which is implemented here. The approach is to change the region in
Mapping's RegionAndPathController, this triggers the plotted region to
shift to the new position, and the trace event triggers the update to
the Acquisition.

A countdown latch is used to delay the running of the acquisition until
the final event above occurs.

Also includes some general tidying of the code (mostly removed unused
code).

Change-Id: Ib99c7d89470e3efb1b2d155b48d9a7ac0933e6dd