I08-374 Set id energy-gap mapping before & after polarisation scan
Mappings are different for linear & circular scans, so set
appropriately.

As these mappings cannot be changed in dummy mode, the set_mapping()
function prints the caput command to the Jython console instead for
debugging purposes.

The mappings are set back to linear in restore_polarisation.py, and the
ScanRequest is configured so this script will always be run, even if the
scan fails.

If in doubt, set_mappings_for_linear_polarisation() in
id_energy_gap_mapping.py can be run on the command line.

Change-Id: I91fdd63fc99227ebc32870cf8d1daf7fc39f4ce2