I20_1-344 : TurboXasScan updates after beamline tests
- Detector readout loop now waits for pulses to be captured by the
Zebra(s) before starting. This avoids a problem where the readout loop
would sometimes end prematurely if no frames of data are available as
soon as or shortly after it starts.
- Zebra(s) are now configured once at the beginning of the scan,
rather than at the start of each collection of spectra.
- Reset number of captured pulses to zero when configuring Zebra(s).
- Improved 'Info' level logging to make progress of the scan easier to
follow.
- Poll interval is set from gda.scannablebase.polltime.millis property
(i.e. same method as used in ScannableBase).

Change-Id: I0896bd0ee8527bf0b289936278de6585f90f46bb