I20_1-348 : Allow experiment setup GUIs to use XML from scan Nexus files
Nexus files from Ede and TurboXas scans contain an XML bean with
a complete description of the experiment, the same as can be saved
from and loaded into experiment setup GUIs. This change is to allow
XML settings to be loaded into the GUI from the scan Nexus files, as
well as from the ASCII XML files.

Change-Id: I79f748c1f39e10ec141eb52bba577c8b64a86d0d