I20-535 : Updates for XES spectrometer
- ScannableGroupAllowedToMove has been refactored to a separate class,
XesSpectrometerCrystal. This extension of ScannableGroup contains
references to the scannables to control x, y, pitch, and rotation,
and an index to identify the horizontal position of the of crystal
in the spectrometer.
- Crystal scannables are injected into XesSpectrometer using the new
XesSpectrometerCrystal groups rather than as individual scannables.
- A new unit test has been added to check positions of the spectrometer
scannables match the previous results for a range of Bragg angles.
Results were generated using GDA 9.22.

Change-Id: I2a1ec920c4fee9f72b91cd892a96716d1386d180