K11: Add Beam Selector Scan functionality
This change adds a UI section allowing users to define and submit a so-
called beam selector scan. The scan is performed by one of two Malcolm
scans (selected by choosing the type of imaging beam desired). Malcolm
does not move the mapping axes at all, so the section lets the users
position them and injects these positions into the start position (which
is handled by GDA).

Save and load are not yet supported.

Change-Id: I665a427afddd11be93eddb0786630bef51b1ea8d