I20_1-357 : Check if EdeDetector is configured before updating GUI
Changes to help avoid the GUI from locking up when using the XH detector
following a server restart. This was caused by a slow Xh 'configure'
method getting called when getting the number of accumulations in order
to generate the GUI,
- The number of accumulations in the TimingGroup section of the 'Time
resolved experiment' GUI will now only be updated  if the currently
selected detector has been configured.
- If the detector has not yet been configured, a dialog is displayed to
allow the user to configure it using the ConnectDeviceDialog (same as
can be done from the 'Beamline Alignment' view).

Change-Id: I4fd62a75e97e973ba53cd6c02962aaa31820e968