Refactor diffraction and point and shoot views
Scans are now defined via intuitive GUIs which are easier to support.
Load listeners, though still not perfect, have a more controlled life
cycle, being attached to each view's composite when it gets created and
disposed when the composite is disposed. This fixes plotting issues
where the path inside a region stops tracking the region.

Change-Id: Ie71aa9a5e21b8677a6bb04a8ec3b4efaf8b45031