I15-694 (9.23) Fix typo in user_commands.py
This was causing an error when the beamline switched to any detector
other than pil3cbf.

Also fixed the name property of the pil3hdf detector so that it didn't
select the non pil3cbf path through the code, and added a few more trace
logs to help see what the code is doing.

Finally, add copies of the final Zebra configuration files after getting
this to work. See https://confluence.diamond.ac.uk/x/NoaQC

Change-Id: I0fd4bdda762afa5e0302010aee1dfc93f61f987d