I20_1-358 : Speed up Ede scans
- The data writer to use for a scan now uses DefaultDataWriterFactory,
so that it is configurable by gda.data.scan.datawriter.dataFormat
property.
- EdeDetector.maxNumFramesToRead parameter is used to control the
maximum
number of frames to read from a detector in a single operation during
a scan.
- EdeScan objects now only store the lightIt scan data points that are
broadcast to the client (i.e. set by the 'number of sec per spectrum to
publish' parameter). This also reduces the Jython console traffic and
helps improve scan speed.
- inputName of Xh/Xstrip detector is now set to an empty array (the same
as for the Frelon detector). This avoids the 'Dubious detector xh with
input names.' warning).
- Extra logging has been added during detector readout to help monitor
how long different parts of the data collection take during a scan.

Change-Id: Ic0a3172c8b78f174a88907b89568337474c8f8da