DAQ-3911 Move remaining mt scripts into i07
A subset of these scripts are actually used on i07. Moving like this
does not require any of the calling scripts to change as the module
names are at the same level in the namespace.

In the future these could be consolidated with the copies of these
scripts which were taken for i06 and i10 when they were moved out of
gda-mt.

Change-Id: I4798d7228e04828fa8f89b2054cee8c08b1bc70f