I11-611 Prevent unnecessary moves for delta and tth
When using the (now deprecated) move command to move either delta
or tth, the other would be parked. With the new devices automatically
parking other devices, this lead to both motors needing to be parked
before moving. When the same motor is being moved each time, this
meant a lot of unnecessary moves.

Change-Id: I8f4f9574b311f622426af527c82154989ff19e66