K11-1491 Fix incorrect property value which causes NullPointerException
The ticket describes how tomography scans cannot be submitted. This is
true only for those scans with dark and/or flat fields. An unhandled
runtime exception when processing scannable positions for these scans
interrupted the submission process.

AcquisitionPropertyType.BEAM_SELECTOR has recently been (rightly) del-
eted, and this property reference was left over. The effect is that
the type field in this object was never set and a stream operation
dereferenced it, producing a NPE. In this change the property is changed
to DIFFRACTION, fixing the problem.

Change-Id: I5ecdd9aac544f2f755db83cc6a366de48a582538