I16-496 (9.25pre) Test GDA File writing Service
Add a before_scan (BeforeScanSnapshotWriter) bean with a set of
additionalScannableNames corresponding to metadata scannables used by
NexusDataWriter.

Add before_scan to the additionalDeviceNames set in the
CommonBeamlineDevicesConfiguration bean.

Remove previous test .dat files which are no longer useful.

Add new eta test .dat files along with .nxs.tree files and (where
possible) .nsx.nxls files.

Add a Notes file with details of the circumstances in which the tests
were run and a summary of the results.

Add a command to command to update numtracker

Change-Id: If32b0a6ce6bd9baae037769d59bbb08a6f48cc1d