K11-1477 Implement DCM roll and pitch as standard Scannable(Epics)Motor
Original EPICS implementation did not conform to the EpicsMotor model,
but that has since changed. Hopefully this change solves the issue in
the ticket. The previous implementation in GDA used a PV.putWait in the
rawAsynchronousMoveTo impl (possibly a mistake) which may be the source
of the problem. Either way, the more standard scannable implementation
is an improvement.

Change-Id: If8b89566c1a1f463bf75c267208f1458252c98dc