I07-457 Move DIFF1 motors into DeferredScannableGroup
The motors are now in a coordinate system which prevents moving two
motors at the same time via their raw motors.

Instead use the deferred move mechanism which moves all requested
axes together such that their move time is the same.

The wrapped group members are extracted so that they can be referenced
as before.

Change-Id: I1d62868e8d2dd7526aacd864755678ae8103eff9