I18-532 Add CompoundNexusContextAppender for Excalibur.
Two appenders are added inside a compound appender and registered for
Excalibur (or the standard Malcolm diffraction scan in dummy mode). Note
that instead of setting their names to the name of the device, they get
unique names in order to retrieve them via the NexusDeviceService later
(complete hack).

Additionally, a new Jython class DiffractionAppenderManager is instan-
tiated in localStation to allow setting the external files for these
appenders.

Change-Id: Ia4e927c835b10e77d6dd4c25da894b8cf9b8aa89