Remove references to spycat.properties from dummy mode on I08-1.
When starting in live mode, GDA should connect to spycat using the
credentials from a file on /dls_sw. I08-1 is also configured to do this
in dummy mode, which seems unnecessary and makes dummy mode depend on
the shared filesystem. We may never know why since Anthony has left, but
if anyone knows of good reasons, let me know.

Change-Id: Ib41ebd8e4ae0fcbdb77de14f709d1a12b50441ee