K11-1676 Do not keep stale scan composites
NamedCompositeFactory implementations e.g. DiffractionComposite would
create their scan controls once and cache the instance, and so when
createComposite was called again, scanControls would be a stale instance
with reference to disposed controls which new listeners would try to
access. This is now fixed by renaming getScanControls() methods to
createScanControls() which now create a new instance on each invocation.

Consequently the early return when controls are disposed checks in
reload methods can now be removed.

Change-Id: I2d7ca8712e753bf2e52f28b5e2f0f68b53d3024f