I20_1-370 : Workaround for memory clearing issue with XH in Ede scans.
If collection using XH detector is stopped early during a Linear/Cyclic
Ede scan, the memory appears to be in a not empty state for the next
measurement. e.g. if a scan is 'smart stopped', the final 'I0' counts
appear to have counts from final frame of 'It' collection added to them.
Performing a successful collection seems to clear memory, so a short
single frame collection is now done before each non 'light It'
measurement as a precaution.

Change-Id: Ia516c00b012c69f559fb001872c0d1107e97a1a8