DAQ-4141 Use separate directories for client product config
Since 2022-03 the clients have been attempting to run update tasks.
This creates a lock file in a 'p2' directory in the configuration
parent directory.

If a different user runs a client on the same workstation there can be
a chance that the Update Job internal error appears.

Using a subdirectory per user should prevent this.

Change-Id: Ia77385a088491711f8b94ff0dc2b48d62e016616