DAQ-4195 : Replace XStream with Jackson XML for TimeResolvedExperimentParameters
- Annotations added to TimeResolvedExperimentParameters and related classes
to allow XML serialization/deserialization using Jackson.
- Minor changes to accommodate exception handling during serialization.
- XmlMapper serialization classes have been made more generic, to allow
the key and value field names to be specified. Each key value pair can also
optionally be enclosed in an entry field. The object value is stored
as a String or Double depending on whether conversion is possible.
- Unit tests have been added. These verify that several XML parameter files
from the beamline can be successfully deserialized and converted back
to XML, producing same output as the original file.

Change-Id: Ia24efaf64f9394290281fec2ad5f1733d9c8c1da