I14-966 Add the "sparse scanning" XANES acquisition to GDA
- Adds sparse_xanes_scan_detector bean to both client and dummy to
differentiate the type of detector used in both modes. The client one
references "Simulated motors for mapping" and the live one references
the 'XRF scanning', which is the detector they want to use in I14 for
Sparse Xanes scanning

- The mapping view configuration now defines two types of
XanesSubmitScanSection which are differentiated by their bean id. Both
have the usual configuration of the XanesSubmitScanSection plus four
more types of configured parameters:
	- scriptFilePath: before this was set in the class, now because there
are two types of xanes scanning, the script file is set here
	- submitButtonColour: before this was set in the class, now because
there are two types of xanes sacanning, the colours are set here
	- sparseScanning: boolean type to differentiate the type of xanes scan
to hide/visualise some specific controls
	- detectorName: this is only set for sparse xanes scanning as they want
a specific detector to be selected in the GUI, and this references to
the sparse_xanes_scan_detector

- Comments out persistence configuration, this was making gda log too
many messages about mongodb not being connected

- 'Sparse xanes scanning' scripts (sparse_xanes_scan.py) have been
adapted to run with the parameters set by GDA instead of defining their
own detector configuration, region, path, etc

- 'submit_sparse_xanes_scan.py' follows patterns of other types of scans
to run the sparse_xanes_scan.py

- 'sparse_xanes_scan_utilities' defines methods that are common to other
xanes_scans which will help to refactor in the future some of these
scripts, and this also makes the try block more readable

Change-Id: I3602200728fdd258ab8f4b3b8f9d11d7a2bdc455