K11-1703 Merge updated axes to preserve any other axis
If we have a scan document and have updated the x and y axes, instead of
replacing the axes docs completely, we merge the updated x and y into
whatever was already there. If there is a third (or fourth, fifth...)
axis that has not been updated, it will still be there.

Change-Id: Iad71bac194d4f244d2a683ff01012ca0cfcd82d5