I14 Reduce timing outs for connecting to ion chambers
When gda tries to connect to the ion chambers and fails, there is a
timeout of 5 seconds.

This call to the ion chambers happens when they move the dcm_enrg, the
shutters are closed or opened and the ring current changes its value.
This results in many timesouts which makes gda very unresponsive.

This change reduces the timeouts by not trying to get the intensity
position with every update of the "readiness status".

When initialising BeamlineReadinessDisplay, logCurrentValues is called
which tries to get the intensity and if it gets a DeviceException, the
boolean variable ionChambersOff is set to true. Next time there is an
update on the other parameters, there is a check in setReadinessStatus
to set the state as "Unknown" and it doesn't attempt to connect to the
ion chambers.

When the ion chambers' position changes, it calls
handleIntensityUpdate() and updates whether the ion chambers are off,
and calls handleUpdate().

Change-Id: I81d5e4e7a7cdc6647a052ef12fc8ae736a03ab16