B16-378 Add processing detector for dummy mode
Configure the simulated area detector to wrap to a
ProcessingDetectorWrapper which is the type that the beamline uses for
most of its detectors. This should allow for easier replication of
issues and for testing changes.

To reproduce B16-378 in dummy mode:
 * Start server
 * Start client
 * Run disable_nexus()
 * Run scan x 0 5 1 simad 1 simadintensity2d
 * This scan succeeds - can also open simad plot view to see image
   (if this doesn't succeed then the simulator needs to be initialised
   by sending frames to the hdf and tiff plugins)
 * Run enable_nexus()
 * Repeat scan as above
 * This fails with GDA waiting for the third image but AD only written
   two

The expected cause is that DetectorDataProcessor's getPosition method is
being called by Nexus writing where it wasn't for dat files.

Change-Id: Iec279644d86e3e7beab08cc879f5ffaa9df5d119