K11: Remove unnecessary properties and tidy up
A lot of the AcquisitionConfigurationProperties have been superseded by
AcquisitionTemplate objects instantiated in the standard GDA way. Others
are just useless. This change removes:

 - client.acquisitions[i].name
 - client.acquisitions[i].engine
 - client.acquisitions[i].templates[j].defaultTemplate
 - client.acquisitions[i].templates[j].tracks[k]
 - client.acquisitions[i].nexusNodeCopyAppender

The remaining AcquisitionConfigurationProperties are 'cameras',
'startPosition' and 'endPosition' which can be obtained from
AcquisitionTemplate objects and will be addressed in an upcoming change.

Deleted also CameraConfigurationProperties.cameraToBeamMap which is no
longer used.

Change-Id: Ie56ae899e6067aa572b05c5a24dd18d66df7dee6