I11-659 Handle detector position in stageManagement script
When stages are parked or engaged, the order devices are moved depends
on the initial state of the stages (detector first or stages first).
This splits the stage moves up so that stages that need to be parked are
moved first, then the detector is moved into position, then the
remaining stages to be moved are engaged.

This is a slightly slower procedure in some cases as the
parking/engaging can't run at the same time but it simplifies the logic
in the scripts where it is used.

Change-Id: Ifa87f118da25094c560dd71d6162063da65a4214