I20-570 : Updates for Synoptic GUI
- LiveControlsImage can now highlight the colour of the text label
when a scannable is busy. The colour used for highlighted state is set
using the 'highlightColour' property of the LiveControlsImage.
- LiveControlsScannableGroup can accept a ScannableGroup where each
member of the group is a ScannableGroup. The label used for each group
is set using the 'descriptions' field.
- The Synoptic perspective has been updated to show the controls for the
lower set of analyser crystals
- SynopticView can scale the background image. This is controlled by
setting the 'imageScaleFactor' property of SynopticViewConfiguration.

Change-Id: I51fdb3cc9699ef9412f899589460f482cf3a6847