I20-570 : Add controls for upper and lower rows of XES spectrometer
- Client side spring config added to create Live controls views for
analyser and detector positions, crystal cuts, simulated position,
energy calibration and offsets.
- A findable SynopticPerspectiveConfiguration object has been added for
the lower and upper rows of the spectrometer. These contain the names of
the views to be added to the Synoptic perspective.
- SynopticPerspective now implements the IExecutableExtension interface,
allowing the name of a configuration object to be passed in from the
perspective extension point. Synoptic perspectives for the upper and
lower rows of the spectrometer can be generated from the same
perspective factory.
- Synoptic view now expects only the name of the view configuration
object to be passed to it. If no configuration object has been set, a
list of all available views is shown to allow the user to select one
to be opened.
- Configuration for old synoptic views has been removed.

Change-Id: Ia1523a2a5601128554c16e79c3eb60013a52af50