I16-610 (9.28) Override AsyncMonitor & AsyncScannable stop() function
The default stop() implementation was failing as it was trying to move
back to the start position, so it needs to be overridden to prevent
this, otherwise the stats at the end of the scan won't be calculated.

Also, do the same in AsyncMoitor, just to be sure.

scan x 1 5 1 asyncScannable(eta, 10) pil 1]

Change-Id: Id00da827a71c996cd50781e285c6e08613ea7db4