I20-569 : Script for assisting alignment of XES spectrometer analysers
Initial version of script to help align the XES spectrometer analysers.
This has been tested in dummy mode by using a MoveableImageDetector to
produce a 2d image with a bright spot with location controlled by yaw
and pitch motor positions and offsets. This behaves in a similar
way to the diffracted fluorescence image on the medipix detector on the
beamline.

Change-Id: If15eb16c41cfcbe31afd18124076c3cb776686cb