I16-629 (9.28) Default to NexusDataWriter profile for transient launcher
Switch to using !NexusScanDataWriter & NexusScanDataWriter profiles
rather than NexusDataWriter & !NexusDataWriter profiles so the beamline
defaults to NexusDataWriter when neither profile is specified.

This is required when launching the beamline with transient devices,
since these profiles are hidden from the transient launcher (the value
has to match the gda.data.scan.datawriter.dataFormat property).

Change-Id: Ibbb39d1ff7c4c78019f939a0343e7513a8abf56b