K11-1735 Toggle Ar/He gas valves automatically as eh_shutter is moved
eh_shutter changed to be a JythonEnumPositionerWrapper, wrapping an
instance of GasSavingShutter. This has reference to an EpicsValve, the
'search and locked' interlock, and the relevant gas valves with their
own interlock. When closing the shutter, the shutter itself is closed,
then the gas valves (if gas saving is enabled in EPICS). When opening,
(if enabled) the gas valves are opened, their interlock is reset (if
no other latched bits are found), then the shutter is opened. There may
be a delay before resetting the gas valves' interlock determined by a
PV search using the EPICS archiver and search parameters read from other
PVs. Dummy components are also included for testing.

On the client, ShutterWidget now schedules an Eclipse Job for toggling
the shutter. The job attaches an observer on the scannable which takes
the message of any SimpleCommandProgress as a subtask, for the sake of
better feedback to the user. Because the job (especially opening) could
take a few minutes, the toggle button now has transition states (Opening
and Closing) with a different colour scheme. These are selected by an
enum status observer. Additionally, for steady state statuses (IDLE,
ERROR), the button position is synchronised with the scannable position.
Finally, only the 'searched and locked' interlock is now used to enable/
disable the shutter control. This is because 1) the 'non critical' has
the gas valves and is now reset internally, and 2) the 'guarlines' was
incorrectly implemented but is a subset of 'search and locked' anyway.

Change-Id: I23d8e68b35acdda1af6cd5f014a80b8a2db51c54