I16 Update scan archiving and dummy mode operation procedures
Move the "Preparing a scan for archiving here" procedure to it's own
file, and update it to rename nxs text files according to whether the
nexus file was written by NexusDataWriter or NexusScanDataWriter. This
should make it easier to see which fie are which, at a glance, rather
than having to did down into the nexis nodes.

Add a script for initialising a basic test ub matrix and update dummy
mode operation notes to run it. This should make it easier to compare
future scans, as they will have more things in common.

Add .nxs and .hdf5 files to .gitignore to ensure they aren't committed
by accident.

Add a template for notes in new test folders.

Change-Id: I989b5f273818ac6f14e04474f70f4ffed8112d4a