I16 Improve test notes and dummy mode startup and scan speed
Only import TripodToolBase if it's needed.

Add DummyScannables in dummy mode for Epics PV scannables used in tests
to remove the pauses due to waiting for PVs to time out.

Set an idgap position in dummy mode to prevent lots of reports of idgap
being outside the range of the lookup table.

Change-Id: Ia921e7086ad6e5790df8455a506a359b7dac33fa