B18-531 : Add new scannables and live control for Meca robot
- New scannable to move Meca robot sample plate to/from the scanning
area (MecaSampleHolderControl). Robot and gripper status words are used
to check current state of the robot (e.g. moving, error etc), so that
moveTo blocks until the move has complete and exception is thrown if the
robot goes into and error state.
- New MecaRobotMover scannable can be used to move one of the
angles/pose positions of the robot at a time (e.g. in a step scan). It
performs the correct sequence of Epics record updates to achieve this :
copy latest readback positions to demand positions, update demand value
for angle/pose to be moved, activate the move.
- New live control 'bit field' - to show the bit representation of the
number returned from a scannable as a series of coloured boxes.

Change-Id: I4af11641715dc41a551ca0e3f002628bafa7b8dc