B16-378 Return cached data from DetectorDataProcessor for Nexus writing
getPosition() is only called on these objects by the Nexus writer
to initialise datasets. As implemented in DetectorDataProcessorWithRoi's
class hierarchy, getPosition() creates a new position callable, which in
this case ends up advancing the state in the TIFF writer and breaking
the scan. The new wrapper returns with either a previous call() result,
or an array (dummy data) of the correct size.

Change-Id: I34091d89405bc286767562b9b998eec732d856c7