I20-591 : Updates to add 2-colour XES scans
XES scans now supports scanning the spectrometer using two different
energies simultaneously. This is achieved by using ScannableGroup
containing upper and lower energy scannables, along with a ScanPositionProvider
that returns an array containing the two energies at each point in the scan.
- XesScanPositionProvider is used to provides the energy value(s) to be
used in the scan. It contains a 'primary' list of energy values (defined
by start, stop, step parameters), and 'secondary' list set using start and step
with number of points equal to that in the primary list. This returns
either a single value (if only primary list has been set), or an array
of 2 values (if both lists have been set).
- Xes scans that move the spectrometer now use XesScanPositionProvider as
the source of positions for the spectrometer scannable. Two-colour scans
use a ScannableGroup (containing upper and lower energy scannables) and
a position provider that has primary and secondary energy lists set.

Change-Id: I255c79cefff969dba1536aa613c904b6ae09c779