I15 (9.30) Tidy localStation and improve dummy mode
Suppress initialisation which requires PV access when in dummy mode.

Suppress check_zebra call when in dummy mode.

Cope better with exposeDetector not being defined, if localStationUser
and/or localStationStaff fail.

Remove references to unused localStationConfiguration parameters and
move all localStationConfiguration imports into a single group at the
top of localStation.

MOve some other imports to the top of localStation.

Change-Id: I2bd89e299a955ece837e086f65b00f3aaaa99431