I15 (9.30) Remove suggestion to replace NDFileImpl with a factory bean
This only works with a normal NDFile plugin, for a camserver NDFile,
there is no plugin base, so an "'ndPluginBase' is not declared" warning
will always be generated. Fixing this will either require a
CamserverPluginBase class to be created or a change to NDFile.

Since the mar also has the same problem, refer back to the comment on
pil3_camserver_ndfile too.

Change-Id: I916e691a96795a2351d3de8daf7b3c1472cbe8fe