I20-589 : Add 'ion chamber readout' checker
A new TestIonchamberReadout Jython scannable has been created that can
be used to test whether the output from ionchambers is consistent.
The scannable is not intended to be used in scans; rather, the test
can be run by calling atScanStart or runCheck methods.
A text string describing the the result from the last time the
checker was run is returned by the getPosition() method.
This 'status' position is added to the Ascii header and Nexus 'before
scan' metadata.

Change-Id: Iba2784d0bd4604a92596a7daa4b2afb8aaa03cf0