I15 (9.30) Move Perkin Elmer detector into a profile
The PE detector is unlikely to be used in the near future.

Currently GDA clients can't access server profile info, so there is
currently no way to re-enable the client side PE spring config if this
is selected as a transient device when launching the server.

Remove the Perkin Elmer TwoDArrayView and MJPegView plugins and menu
contributions, as these have been deprecated for a long time.

Change-Id: I759c9d10eda4a1cab91c8325154c41ecda12e566